opt: SEQUOIADBMAINSTREAM-3855 编目自动切分算法分配不均匀
git-svn-id: http://192.168.20.11/sequoiadb/trunk@37102 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cat/catCommon.cpp
@@ -4607,7 +4607,7 @@ namespace engine
             const std::string &grpName = iterGrp->first ;
             BSONObj lowBound, upBound ;
 
-            if ( itemID >= modMark )
+            if ( (UINT32)itemID >= modMark )
             {
                ++endBound ;
             }
